Move StaffView to separate page file - fix unused imports
- Extract StaffView component to pages/StaffView.jsx
- App.js now only 36 lines (down from 243)
- All unused imports removed
- Fully lazy-loaded with code splitting
- Cleaner separation of concerns
